Tweaks & Audit patch update.

+++ data/scripts/entity/story/aibehaviour.lua
@@ -29,7 +29,7 @@ end
 
 
 function getUpdateInterval()
-    time = time or random():getInt() * 0.5 + 0.5
+    time = time or random():getFloat() * 0.5 + 0.5
     return time
 end
 
@@ -99,7 +99,7 @@ function onBreak(entityId, ...)
             local numTurrets = newPlan.numBlocks / 25 + 1
             AI.addTurrets(child, numTurrets)
 
-            child.shieldDurability = child.shieldMaxDurability * random():getInt()
+            child.shieldDurability = child.shieldMaxDurability * random():getFloat()
 
             WreckageCreator(child.index).active = false
             child:invokeFunction("story/aibehaviour.lua", "setAngry")
