+++ data/scripts/entity/story/xsotanbehaviour.lua
@@ -38,7 +38,7 @@ function XsotanBehaviour.despawnSoon()
     if Galaxy():sectorInRift(x, y) then return end
 
     provoked = false
-    deferredCallback(60 + random():getInt() * 4, "tryDespawn")
+    deferredCallback(60 + random():getFloat() * 4, "tryDespawn")
 end
 
 function XsotanBehaviour.tryDespawn()
